Handle better misconfiguration
Many users have problem trying to setup a benchmark with interpolations.
A common use case is:

```yaml
- name: Fetch user
  request:
    url: {{ memory.body.remote_url }}
```

This is an invalid payload because those curly brackets are valid YAML
syntax for hashes and drill expects a string there. To solve this, the
user needs wrap interpolation with quotes.

This commit improves the shown message when this happens.